% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BayesGO-class}
\alias{BayesGO-class}
\title{An S4 class to represent bayesGO model fitting results.}
\description{
An S4 class to represent bayesGO model fitting results.
}
\section{Slots}{

\describe{
\item{\code{data}}{data}

\item{\code{param}}{model parameter settings}

\item{\code{init}}{model initialization}

\item{\code{fit}}{rjags fitting results}

\item{\code{out}}{clustering and association analysis results}
}}

