% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesGO.R
\name{bayesGO}
\alias{bayesGO}
\title{bayesGO function}
\usage{
bayesGO(pmat, nRow = NA, nCol = NA, Vmax = NA, Kmax = NA,
  nBurnin = 20000, nMain = 10000, thin = 10, seed = 12345)
}
\arguments{
\item{pmat}{A matrix of hypergeometric p-values}

\item{nRow}{Number of rows to analyze (default: use all rows)}

\item{nCol}{Number of columns to analyze (default: use all columns)}

\item{Vmax}{Maximum number of row clusters (default: 0.1 * nRow)}

\item{Kmax}{Maximum number of column clusters (default: 0.1 * nCol)}

\item{nBurnin}{MCMC: Number of burn-in iterations (default: 20000)}

\item{nMain}{MCMC: Number of main iterations (default: 10000)}

\item{thin}{Thinning (default: 10)}

\item{seed}{Random seed (default: 12345)}
}
\description{
Fit the bayesGO model
}
\examples{
data(simul)
fit.bayesGO <- bayesGO( simul, nBurnin=10, nMain=10, thin=1 )
fit.bayesGO
predict( fit.bayesGO, thresGene=0.9, thresTerm=0.9 )
plot( fit.bayesGO, thresGene=0.9, thresTerm=0.9 )
}
