% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BayesGO.R
\docType{methods}
\name{plot,BayesGO,missing-method}
\alias{plot,BayesGO,missing-method}
\alias{plot,BayesGO-method}
\title{Plotting function}
\usage{
\S4method{plot}{BayesGO,missing}(x, y, thresGene = 0.9, thresTerm = 0.9)
}
\arguments{
\item{x}{output of bayesGO function}

\item{y}{(unused argument)}

\item{thresGene}{threshold for high confidence assignment to gene clusters, which is between 0 and 1 (default: 0.9)}

\item{thresTerm}{threshold for high confidence assignment to GO term clusters, which is between 0 and 1}
}
\description{
Plot a heatmap of association between GO terms and genes.
}
