% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmat.R
\docType{data}
\name{pmat}
\alias{pmat}
\title{real data for human genes}
\format{A matrix of hypergeometric p-values, where rows and columns correspond to GO terms and genes.}
\usage{
pmat
}
\description{
real data for human genes
}
\examples{
data(pmat)
}
\keyword{datasets}
