% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BayesGO.R
\docType{methods}
\name{predict,BayesGO-method}
\alias{predict,BayesGO-method}
\title{Gene and GO term cluster prediction}
\usage{
\S4method{predict}{BayesGO}(object, thresGene = 0.9, thresTerm = 0.9)
}
\arguments{
\item{object}{output of bayesGO function}

\item{thresGene}{threshold for high confidence assignment to gene clusters, which is between 0 and 1 (default: 0.9)}

\item{thresTerm}{threshold for high confidence assignment to GO term clusters, which is between 0 and 1}
}
\description{
Return gene and GO term clustering results and their association patterns.
}
