% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BayesGO.R
\docType{methods}
\name{show,BayesGO-method}
\alias{show,BayesGO-method}
\title{Summary of bayesGO results}
\usage{
\S4method{show}{BayesGO}(object)
}
\arguments{
\item{object}{output of bayesGO function}
}
\description{
Summary of bayesGO results.
}
