% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\docType{data}
\name{simul}
\alias{simul}
\title{simulation data}
\format{A matrix of hypergeometric p-values, where rows and columns correspond to GO terms and genes.}
\usage{
simul
}
\description{
simulation data
}
\examples{
data(simul)
}
\keyword{datasets}
