% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{real data for yeast genes}
\format{A matrix of hypergeometric p-values, where rows and columns correspond to GO terms and genes.}
\usage{
yeast
}
\description{
real data for yeast genes
}
\examples{
data(yeast)
}
\keyword{datasets}
